requirements.txt
# Python依赖
requests==2.28.1
croniter==1.3.8
python-dotenv==0.21.0
certifi==2023.7.22
PyYAML==6.0.1